requirements.txt
python-telegram-bot==20.7
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
aiohttp==3.9.1
beautifulsoup4==4.12.3
feedparser==6.0.11
nltk==3.8.1
pandas==2.2.0
